w9-w10-Assignments/requirements.txt
fastapi
uvicorn[standard]
streamlit
requests
python-dotenv
pytest
langgraph
google.generativeai